move SimpleFileImporter to its own module
