+++ tests/core/test_model.py
@@ -8,7 +8,7 @@ import pytest
 import rasa
 import rasa.core
 import rasa.nlu
-from rasa.importers.importer import SimpleFileImporter
+from rasa.importers.simple import SimpleFileImporter
 from rasa import data
 from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DATA_PATH
 from rasa.core.domain import Domain
